requirements.txt
torch
transformers
faiss-cpu
sentence-transformers
streamlit
pandas
numpy
scikit-learn